Исправлены SQL-запросы в сервисе турниров для корректной обработки идентификаторов участников. Обновлено приведение типов для параметров запроса, что улучшает стабильность работы с базой данных. Рекомендуется протестировать функциональность добавления участников турнира на корректность после изменений.

+++ backend/services/tournament/TournamentService.js
@@ -421,7 +421,7 @@ class TournamentService {
                             `SELECT tp.user_id, COALESCE(tp.name, u.username, 'Qualified #' || $3) AS name
                              FROM tournament_participants tp
                              LEFT JOIN users u ON u.id = tp.user_id
-                             WHERE tp.id = $2`,
+                             WHERE tp.id = $2::int`,
                             [finalTournamentId, refId, String(refId)]
                         );
                         const userId = srcPartRes.rows[0]?.user_id || null;
@@ -473,7 +473,7 @@ class TournamentService {
                          SELECT $1, tp.user_id, COALESCE(tp.name, u.username, 'Qualified #' || $3), false
                          FROM tournament_participants tp
                          LEFT JOIN users u ON u.id = tp.user_id
-                         WHERE tp.id = $2
+                         WHERE tp.id = $2::int
                          AND NOT EXISTS (
                             SELECT 1 FROM tournament_participants p
                             WHERE p.tournament_id = $1 AND (p.user_id = tp.user_id OR p.name = COALESCE(tp.name, u.username))
